+++ Lib/Composition/Composition.cs
@@ -21,6 +21,7 @@ using System.Reflection;
 using System.Text;
 using System.Reactive;
 using Lib.BuildCache;
+using Lib.Registry;
 using Lib.Utils.Notification;
 using Lib.Translation;
 using Lib.Utils.Logger;
@@ -88,6 +89,7 @@ namespace Lib.Composition
                 else
                     _buildCache = new PersistentBuildCache(_tools.Path);
             }
+
             if (_command is BuildInteractiveCommand)
             {
                 RunInteractive(_command as CommonInteractiveCommand);
@@ -109,6 +111,55 @@ namespace Lib.Composition
             {
                 RunTranslation(tCommand);
             }
+            else if (_command is PackageManagerInstallCommand installCommand)
+            {
+                RunPackageInstall(installCommand);
+            }
+            else if (_command is PackageManagerUpgradeCommand upgradeCommand)
+            {
+                RunPackageUpgrade(upgradeCommand);
+            }
+            else if (_command is PackageManagerAddCommand addCommand)
+            {
+                RunPackageAdd(addCommand);
+            }
+        }
+
+        void RunPackageAdd(PackageManagerAddCommand addCommand)
+        {
+            InitDiskCache();
+            var project = AddProject(PathUtils.Normalize(Environment.CurrentDirectory), false);
+            var currentNodePackageManager = new CurrentNodePackageManager(_dc, _logger);
+            if (addCommand.PackageName.Value == null)
+            {
+                _logger.Error("Didn't specified name of package to add");
+            }
+            else
+            {
+                currentNodePackageManager.Add(project.Owner.Owner,addCommand.PackageName.Value,addCommand.Dev.Value);
+            }
+        }
+
+        void RunPackageUpgrade(PackageManagerUpgradeCommand upgradeCommand)
+        {
+            InitDiskCache();
+            var project = AddProject(PathUtils.Normalize(Environment.CurrentDirectory), false);
+            var currentNodePackageManager = new CurrentNodePackageManager(_dc, _logger);
+            if (upgradeCommand.PackageName.Value == null)
+            {
+                currentNodePackageManager.UpgradeAll(project.Owner.Owner);
+            }
+            else
+            {
+                currentNodePackageManager.Upgrade(project.Owner.Owner,upgradeCommand.PackageName.Value);
+            }
+        }
+
+        void RunPackageInstall(PackageManagerInstallCommand installCommand)
+        {
+            InitDiskCache();
+            var project = AddProject(PathUtils.Normalize(Environment.CurrentDirectory), false);
+            new CurrentNodePackageManager(_dc, _logger).Install(project.Owner.Owner);
         }
 
         void RunTranslation(TranslationCommand tCommand)
@@ -133,6 +184,7 @@ namespace Lib.Composition
                     trDb.SaveLangDb(PathToTranslations(project), addLanguage);
                     _logger.WriteLine($"Added language {addLanguage}");
                 }
+
                 return;
             }
 
@@ -151,6 +203,7 @@ namespace Lib.Composition
                     File.Delete(PathUtils.Join(PathToTranslations(project), $"{removeLanguage}.json"));
                     _logger.WriteLine($"Removed language {removeLanguage}");
                 }
+
                 return;
             }
 
@@ -208,6 +261,7 @@ namespace Lib.Composition
                         _logger.Error("Import failed. See output for more information.");
                         return;
                     }
+
                     var importedLang = Path.GetFileNameWithoutExtension(PathUtils.Normalize(import));
                     trDb.SaveLangDb(PathToTranslations(project), importedLang);
 
@@ -225,14 +279,15 @@ namespace Lib.Composition
                     var dir = Path.GetDirectoryName(specificPath);
                     trDb.SaveLangDb(dir, language);
 
-                    _logger.WriteLine($"Translated language from file {import} successfully imported to file {specificPath}.");
+                    _logger.WriteLine(
+                        $"Translated language from file {import} successfully imported to file {specificPath}.");
                 }
 
                 return;
             }
 
             var union = tCommand.Union.Value;
-            if (union != null && union.All(x => x!= null))
+            if (union != null && union.All(x => x != null))
             {
                 if (union.Length != 3)
                 {
@@ -243,14 +298,14 @@ namespace Lib.Composition
                 project.InitializeTranslationDb();
                 trDb = project.TranslationDb;
 
-                if(trDb.UnionExportedLanguage(union[0], union[1], union[2]))
+                if (trDb.UnionExportedLanguage(union[0], union[1], union[2]))
                     _logger.Success($"Union of {union[0]} with {union[1]} successfully saved to {union[2]}");
 
                 return;
             }
 
             var subtract = tCommand.Subtract.Value;
-            if (subtract != null && subtract.All(x => x!= null))
+            if (subtract != null && subtract.All(x => x != null))
             {
                 if (subtract.Length != 3)
                 {
@@ -261,8 +316,9 @@ namespace Lib.Composition
                 project.InitializeTranslationDb();
                 trDb = project.TranslationDb;
 
-                if(trDb.SubtractExportedLanguage(subtract[0], subtract[1], subtract[2]))
-                    _logger.Success($"Subtract of {subtract[0]} with {subtract[1]} successfully saved to {subtract[2]}");
+                if (trDb.SubtractExportedLanguage(subtract[0], subtract[1], subtract[2]))
+                    _logger.Success(
+                        $"Subtract of {subtract[0]} with {subtract[1]} successfully saved to {subtract[2]}");
 
                 return;
             }
@@ -276,7 +332,8 @@ namespace Lib.Composition
             AppDomain.CurrentDomain.FirstChanceException += CurrentDomain_FirstChanceException;
         }
 
-        void CurrentDomain_FirstChanceException(object sender, System.Runtime.ExceptionServices.FirstChanceExceptionEventArgs e)
+        void CurrentDomain_FirstChanceException(object sender,
+            System.Runtime.ExceptionServices.FirstChanceExceptionEventArgs e)
         {
             string s = e.Exception.ToString();
             if (s.Contains("KestrelConnectionReset"))
@@ -306,7 +363,8 @@ namespace Lib.Composition
                     proj.Owner.InitializeOnce();
                     proj.OutputSubDir = bCommand.VersionDir.Value;
                     proj.CompressFileNames = !bCommand.Fast.Value;
-                    proj.StyleDefNaming = ParseStyleDefNaming(bCommand.Style.Value ?? (bCommand.Fast.Value ? "2" : "0"));
+                    proj.StyleDefNaming =
+                        ParseStyleDefNaming(bCommand.Style.Value ?? (bCommand.Fast.Value ? "2" : "0"));
                     proj.BundleCss = !bCommand.Fast.Value;
                     proj.Defines["DEBUG"] = bCommand.Fast.Value;
                     proj.GenerateCode();
@@ -325,7 +383,8 @@ namespace Lib.Composition
                     var buildResult = ctx.BuildResult;
                     var filesContent = new Dictionary<string, object>();
                     proj.FillOutputByAdditionalResourcesDirectory(filesContent);
-                    IncludeMessages(buildResult, ref errors, ref warnings, messages, messagesFromFiles, proj.Owner.Owner.FullPath);
+                    IncludeMessages(buildResult, ref errors, ref warnings, messages, messagesFromFiles,
+                        proj.Owner.Owner.FullPath);
                     if (errors == 0)
                     {
                         if (proj.Localize && bCommand.UpdateTranslations.Value)
@@ -350,6 +409,7 @@ namespace Lib.Composition
                                 bundle.BuildResult = buildResult;
                                 bundle.Build(bCommand.Compress.Value, bCommand.Mangle.Value, bCommand.Beautify.Value);
                             }
+
                             SaveFilesContentToDisk(filesContent, bCommand.Dir.Value);
                             totalFiles += filesContent.Count;
                         }
@@ -361,9 +421,13 @@ namespace Lib.Composition
                     errors++;
                 }
             }
+
             var duration = DateTime.UtcNow - start;
-             var color = errors != 0 ? ConsoleColor.Red : warnings != 0 ? ConsoleColor.Yellow : ConsoleColor.Green;
-            _logger.WriteLine("Build done in " + duration.TotalSeconds.ToString("F1", CultureInfo.InvariantCulture) + "s with " + Plural(errors, "error") + " and " + Plural(warnings, "warning") + " and has " + Plural(totalFiles, "file"), color);
+            var color = errors != 0 ? ConsoleColor.Red : warnings != 0 ? ConsoleColor.Yellow : ConsoleColor.Green;
+            _logger.WriteLine(
+                "Build done in " + duration.TotalSeconds.ToString("F1", CultureInfo.InvariantCulture) + "s with " +
+                Plural(errors, "error") + " and " + Plural(warnings, "warning") + " and has " +
+                Plural(totalFiles, "file"), color);
 
             Environment.ExitCode = errors != 0 ? 1 : 0;
         }
@@ -395,6 +459,7 @@ namespace Lib.Composition
             {
                 port = portInInt;
             }
+
             StartWebServer(port, false);
             DateTime start = DateTime.UtcNow;
             int errors = 0;
@@ -407,7 +472,6 @@ namespace Lib.Composition
             {
                 try
                 {
-
                     _logger.WriteLine("Test build started " + proj.Owner.Owner.FullPath, ConsoleColor.Blue);
                     TestResultsHolder testResults = new TestResultsHolder();
                     proj.Owner.LoadProjectJson(true);
@@ -438,7 +502,8 @@ namespace Lib.Composition
                         fastBundle.Build("bb/base", "testbundle.js.map", true);
                         proj.TestProjFastBundle = fastBundle;
                         proj.FilesContent = filesContent;
-                        IncludeMessages(proj.TestProjFastBundle, ref errors, ref warnings, messages, messagesFromFiles, proj.Owner.Owner.FullPath);
+                        IncludeMessages(proj.TestProjFastBundle, ref errors, ref warnings, messages, messagesFromFiles,
+                            proj.Owner.Owner.FullPath);
                         if (errors == 0)
                         {
                             var wait = new Semaphore(0, 1);
@@ -450,16 +515,20 @@ namespace Lib.Composition
                             });
                             var durationb = DateTime.UtcNow - start;
 
-                            _logger.Success("Build successful. Starting Chrome to run tests in " + durationb.TotalSeconds.ToString("F1", CultureInfo.InvariantCulture) + "s");
+                            _logger.Success("Build successful. Starting Chrome to run tests in " +
+                                            durationb.TotalSeconds.ToString("F1", CultureInfo.InvariantCulture) + "s");
 
-                            _testServer.StartTest("/test.html", new Dictionary<string, SourceMap> { { "testbundle.js", testBuildResult.SourceMap } });
+                            _testServer.StartTest("/test.html",
+                                new Dictionary<string, SourceMap> {{"testbundle.js", testBuildResult.SourceMap}});
                             StartChromeTest();
                             wait.WaitOne();
                             StopChromeTest();
                         }
                     }
+
                     if (testCommand.Out.Value != null)
-                        File.WriteAllText(testCommand.Out.Value, testResults.ToJUnitXml(testCommand.FlatTestSuites.Value), new UTF8Encoding(false));
+                        File.WriteAllText(testCommand.Out.Value,
+                            testResults.ToJUnitXml(testCommand.FlatTestSuites.Value), new UTF8Encoding(false));
                 }
                 catch (Exception ex)
                 {
@@ -467,9 +536,14 @@ namespace Lib.Composition
                     errors++;
                 }
             }
+
             var duration = DateTime.UtcNow - start;
-            var color = (errors + testFailures) != 0 ? ConsoleColor.Red : warnings != 0 ? ConsoleColor.Yellow : ConsoleColor.Green;
-            _logger.WriteLine("Test done in " + duration.TotalSeconds.ToString("F1", CultureInfo.InvariantCulture) + " with " + Plural(errors, "error") + " and " + Plural(warnings, "warning") + " and has " + Plural(totalFiles, "file") + " and " + Plural(testFailures, "failure"), color);
+            var color = (errors + testFailures) != 0 ? ConsoleColor.Red :
+                warnings != 0 ? ConsoleColor.Yellow : ConsoleColor.Green;
+            _logger.WriteLine(
+                "Test done in " + duration.TotalSeconds.ToString("F1", CultureInfo.InvariantCulture) + " with " +
+                Plural(errors, "error") + " and " + Plural(warnings, "warning") + " and has " +
+                Plural(totalFiles, "file") + " and " + Plural(testFailures, "failure"), color);
 
             Environment.ExitCode = (errors + testFailures) != 0 ? 1 : 0;
         }
@@ -499,16 +573,17 @@ namespace Lib.Composition
                 var fileName = PathUtils.Join(dir, nameAndContent.Key);
                 if (content is Lazy<object>)
                 {
-                    content = ((Lazy<object>)content).Value;
+                    content = ((Lazy<object>) content).Value;
                 }
+
                 Directory.CreateDirectory(PathUtils.Parent(fileName) ?? ".");
                 if (content is string)
                 {
-                    File.WriteAllText(fileName, (string)content, utf8WithoutBom);
+                    File.WriteAllText(fileName, (string) content, utf8WithoutBom);
                 }
                 else
                 {
-                    File.WriteAllBytes(fileName, (byte[])content);
+                    File.WriteAllBytes(fileName, (byte[]) content);
                 }
             }
         }
@@ -521,6 +596,7 @@ namespace Lib.Composition
             {
                 port = portInInt;
             }
+
             InitDiskCache();
             InitTestServer();
             InitMainServer();
@@ -538,7 +614,8 @@ namespace Lib.Composition
             };
             var location = new Uri(Assembly.GetEntryAssembly().GetName().CodeBase);
             var runningFrom = PathUtils.Normalize(new FileInfo(location.AbsolutePath).Directory.FullName);
-            _bbdir = PathUtils.Join(PathUtils.Normalize(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile)), ".bbcore");
+            _bbdir = PathUtils.Join(
+                PathUtils.Normalize(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile)), ".bbcore");
             if (runningFrom.StartsWith(_bbdir))
             {
                 _bbdir = PathUtils.Join(_bbdir, "tools");
@@ -547,6 +624,7 @@ namespace Lib.Composition
             {
                 _bbdir = PathUtils.Join(_bbdir, "dev");
             }
+
             _tools = new ToolsDir.ToolsDir(_bbdir, _logger);
             _compilerPool = new CompilerPool(_tools, _logger);
             _notificationManager = new NotificationManager();
@@ -571,13 +649,14 @@ namespace Lib.Composition
                 Tools = _tools,
                 BuildCache = _buildCache,
                 Owner = proj,
-                Defines = new Dictionary<string, bool> { { "DEBUG", true } },
+                Defines = new Dictionary<string, bool> {{"DEBUG", true}},
                 SpriteGeneration = enableSpritting
             };
             lock (_projectsLock)
             {
                 _projects.Add(proj.ProjectOptions);
             }
+
             _currentProject = proj.ProjectOptions;
             if (_mainServer != null)
                 _mainServer.Project = _currentProject;
@@ -613,17 +692,20 @@ namespace Lib.Composition
                     context.Response.Redirect("/bb/", true);
                     return;
                 }
+
                 if (bbweb == "/" || bbweb == "/index.html")
                 {
                     await context.Response.WriteAsync(_tools.WebIndexHtml);
                     return;
                 }
+
                 if (bbweb == "/a.js")
                 {
                     await context.Response.WriteAsync(_tools.WebAJs);
                     return;
                 }
             }
+
             if (path.StartsWithSegments("/bb/test", out var bbtest))
             {
                 if (bbtest == "")
@@ -631,32 +713,38 @@ namespace Lib.Composition
                     context.Response.Redirect("/bb/test/", true);
                     return;
                 }
+
                 if (bbtest == "/" || bbtest == "/index.html")
                 {
                     await context.Response.WriteAsync(_tools.WebtIndexHtml);
                     return;
                 }
+
                 if (bbtest == "/a.js")
                 {
                     await context.Response.WriteAsync(_tools.WebtAJs);
                     return;
                 }
             }
+
             if (path == "/bb/api/test")
             {
                 await _testServerLongPollingHandler.Handle(context);
                 return;
             }
+
             if (path == "/bb/api/main")
             {
                 await _mainServerLongPollingHandler.Handle(context);
                 return;
             }
+
             if (path == "/bb/api/projectdirectory")
             {
                 await context.Response.WriteAsync(_currentProject.Owner.Owner.FullPath);
                 return;
             }
+
             if (path.StartsWithSegments("/bb/api/liveReload", out var liveIdx))
             {
                 while (_currentProject.LiveReloadIdx == int.Parse(liveIdx.Value.Substring(1)))
@@ -664,6 +752,7 @@ namespace Lib.Composition
                     await _currentProject.LiveReloadAwaiter.Task;
                     _currentProject.LiveReloadAwaiter = new TaskCompletionSource<Unit>();
                 }
+
                 context.Response.StatusCode = 200;
                 await context.Response.WriteAsync("Reload");
                 return;
@@ -680,6 +769,7 @@ namespace Lib.Composition
                     return;
                 }
             }
+
             var pathWithoutFirstSlash = path.Value.Substring(1);
             var filesContentFromCurrentProjectBuildResult = _currentProject.FilesContent;
             object content;
@@ -688,23 +778,27 @@ namespace Lib.Composition
                 context.Response.ContentType = PathUtils.PathToMimeType(pathWithoutFirstSlash);
                 if (content is Lazy<object>)
                 {
-                    content = ((Lazy<object>)content).Value;
+                    content = ((Lazy<object>) content).Value;
                 }
+
                 if (content is string)
                 {
-                    await context.Response.WriteAsync((string)content);
+                    await context.Response.WriteAsync((string) content);
                 }
                 else
                 {
-                    await context.Response.Body.WriteAsync((byte[])content, 0, ((byte[])content).Length);
+                    await context.Response.Body.WriteAsync((byte[]) content, 0, ((byte[]) content).Length);
                 }
+
                 return;
             }
+
             context.Response.StatusCode = 404;
             await context.Response.WriteAsync("Not found " + path);
         }
 
-        static bool FindInFilesContent(string pathWithoutFirstSlash, Dictionary<string, object> filesContentFromCurrentProjectBuildResult, out object content)
+        static bool FindInFilesContent(string pathWithoutFirstSlash,
+            Dictionary<string, object> filesContentFromCurrentProjectBuildResult, out object content)
         {
             content = null;
             if (filesContentFromCurrentProjectBuildResult == null)
@@ -720,6 +814,7 @@ namespace Lib.Composition
                     return true;
                 }
             }
+
             return false;
         }
 
@@ -729,11 +824,12 @@ namespace Lib.Composition
             _testServerLongPollingHandler = new LongPollingServer(_testServer.NewConnectionHandler);
             _testServer.OnTestResults.Subscribe((results) =>
             {
-              var color = results.TestsFailed != 0 ? ConsoleColor.Red : results.TestsSkipped != 0 ? ConsoleColor.Yellow : ConsoleColor.Green;
+                var color = results.TestsFailed != 0 ? ConsoleColor.Red :
+                    results.TestsSkipped != 0 ? ConsoleColor.Yellow : ConsoleColor.Green;
                 _logger.WriteLine(
-                    $"Tests on {results.UserAgent} Failed: {results.TestsFailed} Skipped: {results.TestsSkipped} Total: {results.TotalTests} Duration: {results.Duration * 0.001:F1}s", color);
+                    $"Tests on {results.UserAgent} Failed: {results.TestsFailed} Skipped: {results.TestsSkipped} Total: {results.TotalTests} Duration: {results.Duration * 0.001:F1}s",
+                    color);
                 _notificationManager.SendNotification(results.ToNotificationParameters());
-
             });
         }
 
@@ -762,11 +858,13 @@ namespace Lib.Composition
                     {
                         toBuild = _projects.ToArray();
                     }
+
                     if (toBuild.Length == 0)
                     {
                         _logger.Error("Change detected, but no project to build");
                         continue;
                     }
+
                     _mainServer.NotifyCompilationStarted();
                     int errors = 0;
                     int warnings = 0;
@@ -805,12 +903,14 @@ namespace Lib.Composition
                             fastBundle.BuildResult = buildResult;
                             fastBundle.Build("bb/base", "bundle.js.map");
                             proj.MainProjFastBundle = fastBundle;
-                            IncludeMessages(proj.MainProjFastBundle, ref errors, ref warnings, messages, messagesFromFiles, proj.Owner.Owner.FullPath);
+                            IncludeMessages(proj.MainProjFastBundle, ref errors, ref warnings, messages,
+                                messagesFromFiles, proj.Owner.Owner.FullPath);
                             if (errors == 0 && proj.LiveReloadEnabled)
                             {
                                 proj.LiveReloadIdx++;
                                 proj.LiveReloadAwaiter.TrySetResult(Unit.Default);
                             }
+
                             if (proj.TestSources != null && proj.TestSources.Count > 0)
                             {
                                 ctx = new BuildCtx(_compilerPool, _verbose, ShowTsVersion);
@@ -828,10 +928,13 @@ namespace Lib.Composition
                                 fastBundle.BuildResult = testBuildResult;
                                 fastBundle.Build("bb/base", "testbundle.js.map", true);
                                 proj.TestProjFastBundle = fastBundle;
-                                IncludeMessages(proj.TestProjFastBundle, ref errors, ref warnings, messages, messagesFromFiles, proj.Owner.Owner.FullPath);
+                                IncludeMessages(proj.TestProjFastBundle, ref errors, ref warnings, messages,
+                                    messagesFromFiles, proj.Owner.Owner.FullPath);
                                 if (errors == 0)
                                 {
-                                    _testServer.StartTest("/test.html", new Dictionary<string, SourceMap> { { "testbundle.js", testBuildResult.SourceMap } });
+                                    _testServer.StartTest("/test.html",
+                                        new Dictionary<string, SourceMap>
+                                            {{"testbundle.js", testBuildResult.SourceMap}});
                                     StartChromeTest();
                                 }
                             }
@@ -839,6 +942,7 @@ namespace Lib.Composition
                             {
                                 proj.TestProjFastBundle = null;
                             }
+
                             proj.FilesContent = filesContent;
                             totalFiles += filesContent.Count;
                         }
@@ -848,11 +952,18 @@ namespace Lib.Composition
                             errors++;
                         }
                     }
+
                     var duration = DateTime.UtcNow - start;
                     _mainServer.NotifyCompilationFinished(errors, warnings, duration.TotalSeconds, messages);
-                    _notificationManager.SendNotification(NotificationParameters.CreateBuildParameters(errors, warnings, duration.TotalSeconds));
-                    var color = errors != 0 ? ConsoleColor.Red : warnings != 0 ? ConsoleColor.Yellow : ConsoleColor.Green;
-                    _logger.WriteLine("Build done in " + (DateTime.UtcNow - start).TotalSeconds.ToString("F1", CultureInfo.InvariantCulture) + "s with " + Plural(errors, "error") + " and " + Plural(warnings, "warning") + " and has " + Plural(totalFiles, "file"), color);
+                    _notificationManager.SendNotification(
+                        NotificationParameters.CreateBuildParameters(errors, warnings, duration.TotalSeconds));
+                    var color = errors != 0 ? ConsoleColor.Red :
+                        warnings != 0 ? ConsoleColor.Yellow : ConsoleColor.Green;
+                    _logger.WriteLine(
+                        "Build done in " +
+                        (DateTime.UtcNow - start).TotalSeconds.ToString("F1", CultureInfo.InvariantCulture) +
+                        "s with " + Plural(errors, "error") + " and " + Plural(warnings, "warning") + " and has " +
+                        Plural(totalFiles, "file"), color);
                     _dc.ResetChange();
                 }
             });
@@ -865,12 +976,14 @@ namespace Lib.Composition
             return number + " " + word + (number > 1 ? "s" : "");
         }
 
-        void IncludeMessages(FastBundleBundler fastBundle, ref int errors, ref int warnings, List<CompilationResultMessage> messages, HashSet<string> messagesFromFiles, string rootPath)
+        void IncludeMessages(FastBundleBundler fastBundle, ref int errors, ref int warnings,
+            List<CompilationResultMessage> messages, HashSet<string> messagesFromFiles, string rootPath)
         {
             IncludeMessages(fastBundle.BuildResult, ref errors, ref warnings, messages, messagesFromFiles, rootPath);
         }
 
-        void IncludeMessages(BuildResult buildResult, ref int errors, ref int warnings, List<CompilationResultMessage> messages, HashSet<string> messagesFromFiles, string rootPath)
+        void IncludeMessages(BuildResult buildResult, ref int errors, ref int warnings,
+            List<CompilationResultMessage> messages, HashSet<string> messagesFromFiles, string rootPath)
         {
             foreach (var pathInfoPair in buildResult.Path2FileInfo)
             {
@@ -893,10 +1006,10 @@ namespace Lib.Composition
                         Text = d.text,
                         Pos = new int[]
                         {
-                            d.startLine+1,
-                            d.startCharacter+1,
-                            d.endLine+1,
-                            d.endCharacter+1
+                            d.startLine + 1,
+                            d.startCharacter + 1,
+                            d.endLine + 1,
+                            d.endCharacter + 1
                         }
                     });
                 }
@@ -910,6 +1023,7 @@ namespace Lib.Composition
                 var chromePath = ChromePathFinder.GetChromePath(new NativeFsAbstraction());
                 _chromeProcessFactory = new ChromeProcessFactory(chromePath);
             }
+
             if (_chromeProcess == null)
             {
                 try
@@ -946,6 +1060,7 @@ namespace Lib.Composition
                 if (line == "q" || line == "quit")
                     break;
             }
+
             ExitWithCleanUp();
         }